+++ packages/create-instance/create-instance.js
@@ -3,14 +3,16 @@
 import { createSlotVNodes } from './create-slot-vnodes'
 import addMocks from './add-mocks'
 import { addEventLogger } from './log-events'
-import { createComponentStubs } from 'shared/stub-components'
-import { throwError, warn, vueVersion } from 'shared/util'
+import { addStubs } from './add-stubs'
+import { throwError, vueVersion } from 'shared/util'
 import { compileTemplate } from 'shared/compile-template'
+import { isRequiredComponent } from 'shared/validators'
 import extractInstanceOptions from './extract-instance-options'
 import createFunctionalComponent from './create-functional-component'
 import { componentNeedsCompiling, isPlainObject } from 'shared/validators'
 import { validateSlots } from './validate-slots'
 import createScopedSlots from './create-scoped-slots'
+import { extendExtendedComponents } from './extend-extended-components'
 
 function compileTemplateForSlots (slots: Object): void {
   Object.keys(slots).forEach(key => {
@@ -32,21 +34,14 @@ export default function createInstance (
   // Remove cached constructor
   delete component._Ctor
 
-  // mounting options are vue-test-utils specific
-  //
   // instance options are options that are passed to the
   // root instance when it's instantiated
-  //
-  // component options are the root components options
-  const componentOptions = typeof component === 'function'
-    ? component.extendOptions
-    : component
-
   const instanceOptions = extractInstanceOptions(options)
 
-  if (options.mocks) {
-    addMocks(options.mocks, _Vue)
-  }
+  addEventLogger(_Vue)
+  addMocks(options.mocks, _Vue)
+  addStubs(component, options.stubs, _Vue)
+
   if (
     (component.options && component.options.functional) ||
     component.functional
@@ -62,62 +57,38 @@ export default function createInstance (
     compileTemplate(component)
   }
 
-  addEventLogger(_Vue)
-
-  const stubComponents = createComponentStubs(
-    // $FlowIgnore
-    component.components,
-    // $FlowIgnore
-    options.stubs
-  )
-  if (options.stubs) {
-    instanceOptions.components = {
-      ...instanceOptions.components,
-      // $FlowIgnore
-      ...stubComponents
+  // Replace globally registered components with components extended
+  // from localVue. This makes sure the beforeMount mixins to add stubs
+  // is applied to globally registered components.
+  // Vue version must be 2.3 or greater, because of a bug resolving
+  // extended constructor options (https://github.com/vuejs/vue/issues/4976)
+  if (vueVersion > 2.2) {
+    for (const c in _Vue.options.components) {
+      if (!isRequiredComponent(c)) {
+        _Vue.component(c, _Vue.extend(_Vue.options.components[c]))
+      }
     }
   }
-  _Vue.mixin({
-    created () {
-      Object.assign(
-        this.$options.components,
-        stubComponents
-      )
-    }
-  })
-  Object.keys(componentOptions.components || {}).forEach(c => {
-    if (
-      componentOptions.components[c].extendOptions &&
-      !instanceOptions.components[c]
-    ) {
-      if (options.logModifiedComponents) {
-        warn(
-          `an extended child component <${c}> has been modified ` +
-          `to ensure it has the correct instance properties. ` +
-          `This means it is not possible to find the component ` +
-          `with a component selector. To find the component, ` +
-          `you must stub it manually using the stubs mounting ` +
-          `option.`
-        )
-      }
-      instanceOptions.components[c] = _Vue.extend(
-        componentOptions.components[c]
-      )
-    }
-  })
+
+  extendExtendedComponents(
+    component,
+    _Vue,
+    options.logModifiedComponents,
+    instanceOptions.components
+  )
 
   if (component.options) {
     component.options._base = _Vue
   }
 
-  const Constructor = vueVersion < 2.3 && typeof component === 'function'
+  // extend component from _Vue to add properties and mixins
+  // extend does not work correctly for sub class components in Vue < 2.2
+  const Constructor = typeof component === 'function' && vueVersion < 2.3
     ? component.extend(instanceOptions)
     : _Vue.extend(component).extend(instanceOptions)
 
-  Object.keys(instanceOptions.components || {}).forEach(key => {
-    Constructor.component(key, instanceOptions.components[key])
-    _Vue.component(key, instanceOptions.components[key])
-  })
+  // Keep reference to component mount was called with
+  Constructor._vueTestUtilsRoot = component
 
   if (options.slots) {
     compileTemplateForSlots(options.slots)
@@ -155,9 +126,13 @@ export default function createInstance (
       Constructor,
       {
         ref: 'vm',
-        props: options.propsData,
         on: options.listeners,
-        attrs: options.attrs,
+        attrs: {
+          ...options.attrs,
+          // pass as attrs so that inheritAttrs works correctly
+          // propsData should take precedence over attrs
+          ...options.propsData
+        },
         scopedSlots
       },
       slots