+++ packages/create-instance/create-slot-vnodes.js
@@ -8,13 +8,12 @@ function createVNodes (
 ): Array<VNode> {
   const el = compileToFunctions(`<div>${slotValue}</div>`)
   const _staticRenderFns = vm._renderProxy.$options.staticRenderFns
-  // version < 2.5
-  if (!vm._renderProxy._staticTrees) {
-    vm._renderProxy._staticTrees = []
-  }
+  const _staticTrees = vm._renderProxy._staticTrees
+  vm._renderProxy._staticTrees = []
   vm._renderProxy.$options.staticRenderFns = el.staticRenderFns
   const vnode = el.render.call(vm._renderProxy, vm.$createElement)
   vm._renderProxy.$options.staticRenderFns = _staticRenderFns
+  vm._renderProxy._staticTrees = _staticTrees
   return vnode.children
 }
 
@@ -22,10 +21,13 @@ function createVNodesForSlot (
   vm: Component,
   slotValue: SlotValue,
   name: string,
-): VNode | string {
+): VNode | Array<VNode> {
   let vnode
   if (typeof slotValue === 'string') {
     const vnodes = createVNodes(vm, slotValue)
+    if (vnodes.length > 1) {
+      return vnodes
+    }
     vnode = vnodes[0]
   } else {
     vnode = vm.$createElement(slotValue)
@@ -41,7 +43,7 @@ function createVNodesForSlot (
 export function createSlotVNodes (
   vm: Component,
   slots: SlotsObject
-): Array<VNode | string> {
+): Array<VNode | Array<VNode>> {
   return Object.keys(slots).reduce((acc, key) => {
     const content = slots[key]
     if (Array.isArray(content)) {